Display FilterMenuHeaderWidget on RC and Watchlist mobile
When AMC is enabled, the FilterMenuHeaderWidget wasn't being added to
the shared filter UI due to a guard clause that checked for mobile.
The expectation of AMC is to have a very similar UI experience to
desktop, so this was confusing to some users.
- Removed some sections where config.isMobile is checked on
MenuSelectWidget
- Added an isMobile flag to FilterMenuHeaderWidget to display
or not display the header buttons depending if the UI is mobile
or not
- Added CSS to improve FilterMenuHeaderWidget's display on mobile

Bug: T356579
Change-Id: If87eb5ead7023a2d53c8b27a0ec2484c026d0cca