tests: BundleSizeTestBase::provideBundleSize should be static (part 1)
Our providers are meant to be static nowadays, see: T332865. Lets
keep more code consistent with that.

This patch introduces a new getBundleSizeConfigData() static method
that will replace the former getBundleSizeConfig() and deprecates it.
The reason for doing this is because since we're calling these methods
in a static method (data provider), and we already have existing overrides
that are dynamic methods, it'll be hard to change without a simple
step by step migration plan.

So, this patch introduces the new static method to be used by child
classes and also deprecates the dynamic method. Extensions or consumers
of the dynamic method will have time to migrate to the static method.

This patch will also take care of late-static binding issues which
will be introduced when calling overrides of child class methods
in the parent class. We're not supposed to use `self::<methodName>()`
here. We should use `static::<methodName>()`.

Also, note that I didn't mark the new method abstract yet because
there are already existing classes that extend the BundleSizeTestBase
class and that will break CI and other consumers. Once all child
classes migrate to the new method, we can mark it as an abstract
method and new classes must implement it.

Bug: T393122
Change-Id: If898e0cf90e7478ccd59a2db9bf41eccf5a5e58c