Add a script to fix wrongly prefixed passwords and remove b/c code
In an old version there was an error in prefixing type B password hashes
as type A password hashes. This was fixed in Icb809274f9f63641. And in
I772de0fb17245d08 some b/c code got added so that the affected
accounts can still log in.

I would like to clean this up since it did cause confusion in the past
on what exactly type A and type B hashes are.

On WMF I fixed this by hand at some point (T112359#10059827) so this
script is only for very old third party wikis.

Change-Id: I142b2c7d9fe88135200561469642ca684eebdad9