actions: Fix handling of redirects to known (non-existing) pages
Before this patch, MediaWiki doesn't handle redirects to know non-
existing pages. It'll just acts as though the redirect query parameter
on the source page is set to `no` and gets stuck on the source page
showing a link to the redirect (so not following it).

But with this patch, this fixes the issue by following redirects to
both existing and non-existing (known) pages.

Change-Id: Icb1cd037565191cc3cd3f659976d09a1b31f20ad