Deprecate `MediaWikiServices::getPerDbNameStatsdDataFactory()`
This feature is gradually loosing its use-case as we migrate our
metrics to prometheus. In prometheus land, we could just use `setLabel()`
and pass in 'wiki' as the key and the current wiki ID as the value.

We will give this 1 release cycle so all consumers of this interface will
upgrade to the new way and then start emiting warnings.

See also: https://wikitech.wikimedia.org/wiki/Graphite/Deprecation_Roadmap

Depends-On: Ib39c7dd0df1628b6ac7cc3e21fbea398b76eee64
Depends-On: I7a61e965f5491f6686569f1192c7ebc5e0d3f262
Change-Id: I50aadbedd4aeb193c69316b03edfffb6ac859b9a