tests: Fix tests implicitly relying on config
These can result in local test failures.

-StatusFormatterTest: clear handlers for the called hooks to avoid
interferences from other extensions (happens with WikimediaMessages's
Hooks class that takes UserOptionsLookup as a constructor parameter).
Also use a no-op mock to fail when a method is called on the mock not
when it is instantiated. Maybe we could check that the method called is
`getOption` with option `language`, but that starts to rely on
implementation details a bit too much, so it's not done here.
-LanguageNameUtilsTestTrait: explicitly set `ExtraLanguageNames` to be
empty when needed. Overrides in LocalSettings can cause test failures
(seen by assigning a name to qqq with the test case "qqq doesn't have a
name").
-DumpMessagesTest: clear MessageCacheFetchOverrides hook handlers to
prevent overrides (seen with WikimediaMessages).
-AvailableRightsTest: acknowledge that LocalSettings can make it fail; I
don't think it can be fixed easily.

Bug: T277470
Change-Id: I10268a0900d57522a84c95de094934b0b80ea118