Fix location of results cache file in group splitting function
In Idd167ef8b1d we introduced the download of PHPUnit results cache
information. One of the patchsets moved the download location to the
system temp folder, but the subsequent load of the file was still
looking in the current working directory.

Fix the location of the results cache file so that the save and load
of the information refer to the same filename.

Bug: T392892
Change-Id: Ibaf36fc35b58be2bc5fbfd6c10663a69574c6589