Only include valid test descriptors in built suites
When building the test suites, there are some virtual test classes
(like LuaEngineTestSkip or SkippedTestCase) for which no corresponding
PHP class on the filesystem exists. Currently the code adds empty
`<file/>` elements to the `phpunit-*.xml` for this classes - this is
not fatal, but confusing.

Only build suites containing test descriptors for which we have
corresponding PHP classes on the filesystem.

Bug: T392892
Change-Id: Ic11eae2ce9d3a75a8bdd91f5024b761f98c8efed