Add ParserOptions::isMessage()
Many callers of ParserOptions::getInterfaceMessage() were using it to
detect whether they are in a message parsing operation, but actually
getInterfaceMessage() reflects Message::$isInterface, which is true
only for messages in the user's language.

So, add ParserOptions::isMessage(), which is always true during message
parsing, regardless of the language.

We originally had OT_MSG for this, but that was merged into
OT_PREPROCESS.

Update OutputPage::parseInternal() which was relying on the interface
flag to suppress the mw-parser-output class. The comment was out of date
since the false branch of setWrapOutputClass() has been hard-deprecated
since 2017.

Bug: T393601
Change-Id: Ib39d5b7fa135efb9679e3bf9349a5f5882c2c7d1