In messages, make {{USERLANGUAGE}} always return the user's language
Some messages need to access the user's current language, even if
wfMessage()->inContentLanguage() is used. So, in the {{USERLANGUAGE}}
core magic word, allow the user language to be returned if we are
parsing a message, even if ParserEnableUserLanguage is false.

In MessageParser, stop setting the user language in ParserOptions. Allow
the request context language to be used. The language passed down from
Message is still used in the output pipeline.

I considered passing the user language down from Message to
ParserOptions, but resolving the user language is slow and fragile,
throwing an exception in early setup and in load.php. ParserOptions
really needs to receive a provider rather than a resolved language.

The way MessageParser was caching ParserOptions was unsafe for
re-entrant parsing, since the object was shared between parser
instances. A child parse could modify options in its parent. So cache
one ParserOptions object per parser in the stack. Benchmarking shows
this change is beneficial for performance.

Did some basic maintenance on MagicVariableTest.

Bug: T393601
Change-Id: Id3635641090bd8d9671eeabc58a4c45ac530879c