tests: Make PHPUnit data provider static in ExtensionJsonTestBase
It is necessary to change getExtensionJson() to static to call it from
static data providers.
Use reflection to detect how to get the value for the extension path to
allow getExtensionJson() to be static.

The data provider provideHookHandlerNames() is overridden in some
extensions, also use reflection to detect how to call it.
The base implementation has own function name
to avoid php fatal when function is redeclared.
The base implementation has own function name
to avoid php fatal when function is redeclared.
Use php magic __call/__callStatic to catch parent calls.

Bug: T393065
Change-Id: Icfe7838844696135c69959599a83bd72cc353a56