jobqueue: Change runJobs.php default from 150M to "3x $wgMemoryLimit"
$wgMemoryLimit defaults to 50M so this is generally a no-op. But this
makes it so that if you had to raise that for other reasons (e.g.
after installing certain extensions, or after raising the max article
size, etc. WMF set this to 1400MiB at the moment), you don't then
find that jobs break because they still use a lower limit by default.

This comes up in MediaWiki support queries in #mediawiki on IRC
and elsewhere.

Avoid this by encoding the default logic so that they go up automatically
together. It remains unchanged that you can set --memory-limit on any
maintenance script invocation to set a lower or higher limit as
desired.

Ref https://wm-bot.wmflabs.org/logs/%23mediawiki/20200501.txt

Change-Id: I1d1a915d257c9d34a49300b8a06846ee11ba6bdd