Add `LOSSLESS` option to `formatnum` parser function
In old MediaWiki versions, {{formatnum:}} was able to format integers of
arbitrary size or floating-point numbers of arbitrary precision. On the
other hand, on current MediaWiki in Wikimedia production, the number
formatting rounds the number, resulting in a loss of precision through
the formatting operation.

Introduce the `LOSSLESS` option to {{formatnum:}} so that precision can
be preserved in contexts where that is important (for example in
Wikibase). The magic word array code (including the static
initialization) is based on pagesincategory() later in the same file.

See also change Iac196f9cb3 for an earlier version of this change.

Bug: T393483
Change-Id: I2372dcec6f890f60efe42acf30db0640dc685117
Co-Authored-By: Arthur Taylor <arthur.taylor@wikimedia.de>