Stats: Add temporary deprecation for addLabel() normalization
Follows-up I9ce0cbf5bbabf8. While the internal exception from addLabel
becomes a a mere PHP Warning in MetricTrait::setLabel, it does have
the side-effect of replacing the object with a NullMetric, which means
the stat is silently dropped as well, and that's a breaking change.

Temporarily restore normalization with a warning, by emitting the
warning directly.

Bug: T378764
Bug: T394053
Change-Id: Ibeb426e8dd6aeb4cc4ef3449694944c3a1d79715