Do not save on Session::renew() when there's nothing to renew
If the session is nearing expiry, we need to save to refresh the
expiry time in the session store. But when it's not, the save is
not useful, and it will result in a pointless extra write if the
session data is dirty.

Bug: T392251
Change-Id: Ib2e39783d466b4408b5535d2f61bf7008340c535
(cherry picked from commit d8f78525cc5579f32e9799c76d9c32fbdde126ed)