Don't save after Session::delaySave() when there's no delayed save
Session::delaySave() creates a scope in which saves will be
delayed until the end of the scope; but it shouldn't save when
there were no save attempts at all within the scope.

This tends to happen especially in
SessionManager::getSessionFromInfo() which is called a lot, e.g.
from WebRequest::getSession(). Even though save() is idempotent,
it will cause a session store write every time the session data
is dirty, which can result in lots of superfluous writes.

Bug: T392251
Change-Id: I0ab7c0eae5b42ac1836eca283d310bed9939837c