Remove TechWish stats that aren't needed any more
These products/features are deployed and stable for a very, very
long time. There is no decision to be made based on these stats.

This was originally added in 2016 via Idc33dd6 and Ia6abe76, meant
to be used in the "expiring watchlist" project. I believe it was
never actually used. The project was finished by the Community Tech
team years later. They never used these stats, or did they?

Bug: T401557
Change-Id: I3edaa44b5f823c44a72d866614177f761ae13689