Replace trivial array_merge with the new [ ... ] syntax
The behavior of the new [ ... ] syntax is the same.

In some cases a much more trivial array operation is available,
like appending or prepending a single value.

The existing test coverage should demonstrate that this doesn't
change any behavior.

There was one actual bug in InterwikiLookupAdapter, I believe.
I hope this was never really called.

Change-Id: I3548cfa63c9b7e608642af7061baaae28ac82e06