watchlist: Allow configurable default watchlist expiry
This adds preferences for default watchlist expiries for:
* Manually watchlisting (watchstar)
* Editing existing pages (watchdefault)
* Creating new pages (watchcreations)
* Rollback (watchrollback)

Doing the same for uploading files, moving pages/files, and deleting
pages/files can be saved for a future patch, if people want these
features.

To err on the side of caution, we only resort to using the preferred
expiry if the page is not already watched. This may be changed in the
future to auto-watch only if the preferred expiry is longer than the
existing one, depending on user feedback.

Bug: T265716
Change-Id: I383b5a9fd276e5d2e7cf679fd8c398bbf1398bb8
Co-authored-by: MusikAnimal <musikanimal@gmail.com>
Co-authored-by: Scardenasmolinar <scardenasmolinar@wikimedia.org>