MergeLogFormatter: Fix `redirect` parameter on wrong link in merge logs
The "redirect=no" parameter was incorrectly being applied to different
links depending on which page's log was being viewed. When viewing the
merge log entry at the source page, it would add redirect=no to the
source link. When viewing at the destination page, it would add it to
the destination link.

This commit ensures that redirect=no is consistently applied to the
source page link in both log entries, regardless of which page's log
is being viewed.

Bug: T395075
Change-Id: I75965930a26c5a83db718683435e68f72dd050a8