Introduce ContentHolderTransformStage
Some passes can be applied to either a DOM document or a serialized
string form of the document. This is for instance the case of the
HandleSectionLinks passes.

This patch proposes a wrapper around both these passes that chooses the
appropriate one. It would be useful in the context of other passes
provided in both text and DOM versions.

Bug: T404459
Change-Id: I30a0659479858118126a01e97706be3eec5c966e