fix(thumbnails): Do not override broken link colors
Broken file thumbnails have `.mw-file-element`s as children of `a.new`s.
To fix their colors:

- set `@background-color-base` instead of `@background-color-base-fixed`
  on thumbs, and set no font color;
- set the previous (fixed) colors on `:not(.mw-broken-media)`.

The latter ensures that non-broken images continue to have the same
colors as they do now (fixed background and font colors). The font color
matters when the image could not load for some reason but there is an
alt text.

The former restores the red color on broken file links while still
ensuring sufficient contrast: the original style (before
609c3055d935fc7cc8ce54230cd4286442f5c60f) resulted in a 2.61 color
contrast in night mode, which is much less than the required (AA-level)
4.5:1; the new style provides a 7.09:1 contrast, which even passes AAA
(7:1). However, this color contrast stands only if the parent element
indeed has a font color that is consistent with `@background-color-base`
(which is set by the skin, not by core), so I can only hope it looks
good on all current and future skins... But this should not make any
difference in non-night mode-aware skins, which AFAIK means everything
but Vector 2022 and Minerva.

Follows-Up: 609c3055d935fc7cc8ce54230cd4286442f5c60f
Bug: T395206
Change-Id: Ieeaf37b52758667c994906527ce6f90e72b9056c