tests: Rework ApiTestCase::expectApiErrorCode
TestCase::runTest changed to private with PHPUnit 10.
To be prepared for the new version move the implementation to the place
where the exception is thrown.

This requires to implement a function catching from callback to support
test cases for ApiBase and ApiMain.

Use a @postCondition to verify the api error code was thrown. This
cannot be done after ApiMain::execute, because it is unknown if there
are one or more requests and maybe a later should fail. The expectation
only verify that the whole test function fails with this api code,
not a specific single request.

Bug: T328919
Depends-On: I9c8e9a1fb1fb1f8d8ba83205ee59c4aeacdbe54d
Change-Id: I79ee8cfcf9e9ddb1baa8482c06920fd45de3ab78