filebackend: Remove redundant "MainWANObjectCache" override
Since 2020 with change Ia60cb0bc18b23, the main cache is enabled in
PHPUnit by default, so this tracking is redundant.

If it needs to override it, it can use setMainCache()
which takes care of everything else automatically, including
WANObjectCache since that is just a wrapper around the main cache.
Although that isn't even needed here.

Change-Id: I43684822548cda8ea5106b7574d95c3b2664b8cd