Skins: Add missing icons to remaining menu items
Menu items define icons that should be associated with menu
items. Previously we only defined the ones that were used by
the Minerva skin with Minerva-specific definitions to fill in
the missing one, but going forward it is better to have these
all defined centrally in core so other skins (namely Vector 2022)
can benefit from them.

Bug: T365631
Depends-On: I2ba38cb59221ea41b6fcf8c1e066f623f12243bf
Depends-On: I2ed43cac6f4ac9d27865f1874a2e588a496022d8
Change-Id: I7b681496e689c22bad1a956e64c12313ae324edb