Use Philippine decimal format for Chavacano
Chavacano de Zamboanga uses Spanish as its fallback language and
inherits its number format from it as a result. Spanish uses
spaces as grouping separators and the comma as the decimal
separator, which is unlike how numbers are formatted in the
Philippines (dots for decimals, commas for groups).

This switches cbk-zam over to using the Philippine number format.

Bug: T395724
Change-Id: I7f98b0121b57c9c2c69d6bc5876b66c0e1f2886b