Skin: skip brightness filter for images with skin-invert-image class
The intention is to not enforce a dim white background for images
in dark mode if class 'skin-invert-image' is explicitly set.
This class name is currently defined by Minerva and Vector skins,
e. g. in file
resources/skins.vector.styles/CSSCustomProperties.less

Some images benefit readability if inverted in dark mode, for
example diagrams or svg formula. Not all do, wikitext needs
to be updated to set the class on a case-by-case basis
(deciding if a dim white background or the inverted image
version is preferrable in dark mode rendering).

-----

Ideally, the attribute selector statements for 'span' and
'figure' would be the same within the function
`dark-mode-image-styles()` .. however, for 'span'-nested images
the css rule was/is conservatively narrowed down to
elements having class `mw-default-size` only.

This decision _presumably_ stems from the observation that
lots of (explicitly scaled) icons and symbols with transparent
areas should render unmodified across light/dark mode
(they ought neither be inverted nor background-filled white).

It is a compromise/trade-off steming from the lack of a
decent discriminator which could sensibly be used here to
tell apart images
* used as icons/symbols (that are ok to be left unstyled) and
* diagrams/graphics (that should be safe-guarded with a default
dim white background, unless classed in a more specific
way such as `skin-invert-image`)

The compromise has some (unscaled) icons/symbols and
some (scaled) diagrams/graphics appear badly,
which is fixable to some degree by changing the wikitext
to be more descriptive on how to render a certain file in
dark mode, setting css class attributes
(.. effectively providing some sort of discriminator).

Bug: T391800
Change-Id: Ib7febb38740051afd061b9ab9223bdcbfe4bff01