Use first class callable syntax
Convert array callables of the form [ $this, 'foo' ] to PHP 8.1 first
class callable syntax $this->foo( ... ), an expression which evaluates
to a Closure.

Accessibility is evaluated when the closure is constructed, instead of
when it is called, which means that some methods don't need to be
public anymore. I used Code Search to evaluate whether methods had any
external callers. For some tests, I just left them as public, since
Code Search went down.

I benchmarked it. The closure is faster to call, but slower to
construct. Construction took about 20ns longer for me, which was
significant in relative terms (70% longer) but it would have to be a
very tight loop for it to make a difference in reality. For comparison,
Parser::handleMagicLinks('') took 950ns.

Change-Id: I6bb987a2fde4b642866bda05f96ce644a9ac6bfc