ResourceLoader: Bump USERJSPARSE_CACHE_VERSION and STORAGE_VERSION
After updating Peast, we found out there is need to bump the cache version. If an invalid ES2016 code was already parsed before the 2017 update, the error is cached. If we don't invalidate the cache by bumping the cache version that error still shows up when same code which is a valid ES2017 is saved again.

For this time, because we already deployed the Peast call change without
bumping the cache version, some edits have already been re-applied that
expect to adopt ES2017 syntax. To avoid users having to manually purge
or re-apply those edits a second time, we also invalidate the browser local storage. When no edit is made on the page, you can otherwise see
the error from the previous time that edit was saved under ES2016. The
STORAGE_VERSION bump, clears the cache and forces a re-download of all stored modules.

Bug: T381537
Change-Id: I319f668023924ffd497667f22328bd1658926452