Gallery: Allow to supply class parameter per individual image
This patch adds the ability to supply a class= parameter to an entry line of a gallery.

If class= is passed to a stand-alone [[File:..]] statement in wikitext, rendered html will have a class attribute added to the wrapping 'span' of the img element.   The patch introduces the same for individual images of a gallery.

The equivalent Parsoid patch is Idf529d8f11f036d454f4ccce39ea16217d9ca491

Bug: T344784
Change-Id: I3b0232eaedb29b6b3b16c30ab950b3fcf646e171