Move Parsoid indicator from top of page to footer
As proposed in T363484#10887966 move the Parsoid indicator from the
top navigation to the footer.

We also introduce a new ParserOutputFlag mirroring the `useParsoid`
property of the ParserOptions.  This is "OR'ed" among ParserOutputs,
like the other flags, meaning that if any part of a combined MCR
ParserOutput was generated with Parsoid, then the merged result will
have this flag set.

This is deliberately kept separate from
PageBundleParserOutputConverter::hasPageBundle() and the
'isParsoidContent' flag in the OutputPipeline, which reflect
properties for the HTML output.  The ParserOutput flag is (for now)
only a UX indicator and does not affect how content is rendered in the
OutputTransform pipeline.

Bug: T363484
Change-Id: I70a19f2e2ae06dc35845d12e541b7cc03174a489