session: Segregate anonymous sessions from authenticated sessions (p2)
This patch adds a class which handles the responsibility of
deciding where to save anonymous sessions to a backend different
from authenticated sessions (logged-in users) using various
indicators. NOTE that for migration purposes, the anonymous store
and authentication store will be set to the same cache type via
configuration in production. This will enable us to exercise the
MultiBackendSessionStore without actually doing the backend split
to different cache types. It'll help us monitor issues and fix them
before we decide to do a proper backend split. Additionally, when
this patch gets merged, this class won't take effect because the
setting for anonymous session cache type is set to `false` by
default which will trigger the SingleBackendSessionStore instead.

As a follow-up to this patch, we will make a config patch to set
the anonymous session cache type `wgSessionCacheType` which today
in production is set to `'kask-session'`.

Why:

- Make an attempt to reduce/spread infrastructure load in charge
  of MediaWiki sessions due to spikes in inorganic traffic to
  authentication related special pages like Userlogin or CreateAccount.
  Visiting these pages performs session writes.
- Visiting the above mention pages performs a session write for
  both a local wiki and the shared domain (for SUL3) and if the
  user doesn't eventually login or create an account (like automated
  bots), this can get problematic because sessions are written
  to the DB and not garbage collected quickly enough.
- We intend to discard anonymous sessions when they're not used
  as fast as possible as they're not extremely important and we
  can always regenerate them if needed.

What:

- Use a different store for anon sessions and authenticated
  sessions.
- Switches stores on the fly during authentication and use
  the appropriate session backend.
- Between a logged-out session and a logged-in session,
  transfer session data to the correct store.

NOTE: Notice that only authenticated sessions are written to
the store in charge of such sessions. And all anonymous sessions
are kept in the configured store set by `wgAnonSessionCacheType`.
This patch also does some logging and monitoring to help during
migration.

Test plan
=========
1. Set `wgAnonSessionCacheType = <what-ever-cache-type>`. Prod
   will be set to `kask-session`.
2. Try any authentication work flow like; create account, login
   or change password.
3. Notice that there should be no change in behavior. Everything
   should work the same as before.
4. Observe in the code that the cache types set for both anon and
   auth session cache type are set properly and session writes or
   reads are happening to and from the corresponding backends.

TODO (as follow-up)
- Write some PHPUnit tests for MultiBackendSessionStore class.
- Add some debug logging for when get()/set()/delete() methods are
  invoked.

Bug: T399194
Bug: T399195
Bug: T402808
Change-Id: I10101c8b928a1236eeb4d338ad16a1cd3db44bed