Clean up exception handling
Add missing `@throws` tag for checked exceptions, as these will be
reported by phan as part of T338426.

In ExternalStoreAccess, narrow a `catch` clause to only catch ES-related
exceptions, not just any exception. This code was originally only
calling `store()` and catching specific DB errors. Then, in
I0be5d9ebecd58b4f it was changed to catch `MWException` instead; in
I76601a86e30f it was updated to just `Exception`; in I9334d36e8bc, the
`isReadOnly` call was added. The code was then moved to
`ExternalStoreAccess` in I40c3b5534fc8. So, there seems to be no good
reason to catch `Exception` here, and in fact, that might be hiding bugs
and problems elsewhere that aren't meant to be hidden.

In JobQueueGroup::getQueueTypes, change checked exception to unchecked.
The doc comments of this and other methods make it pretty clear that the
caller is not supposed to call them on foreign wikis, so the exception
is unchecked. Else we would need to add a lot of `catch`es all over the
place. JobQueueError itself seems problematic as different places seem
to use it as a checked exception, and others as unchecked, but that
isn't addressed in this patch.

In JsonSerializer, the doc update is a follow-up to Ia654359e0fdec3
(JsonDeserializer was already done in I798b422e703d8e).

Change-Id: Ic93fbc8f4564348a8d9bc99cf190dda437f3a595