Back ParserOutput with a ContentHolder rather than rawText+extensiondata
We back ParserOutput with a ContentHolder that replaces both "rawtext"
and "page bundle extension data". The serialization of the ParserOutput
does not change for now; a follow-up patch should clean up that
serialization to serialize the ContentHolder bundle directly rather
than in ExtensionData.

We also take advantage of the opportunity to use some other helper
methods introduced in Parsoid (for example,
Ic990ab108afc2791f9db297a07cfe1bdc59dbbf2) to simplify the code base
and fully support fragments.  We don't use these yet, but they can be
used to hold (eg) indicators, page title, and other bits of HTML
encoded in the ParserOutput.

Bug: T347062
Bug: T348254
Co-Authored-By: C. Scott Ananian <cananian@wikimedia.org>
Co-Authored-By: Isabelle Hurbain-Palatin <ihurbainpalatin@wikimedia.org>
Depends-On: I5f5b0907c49ac58feefc5f26a092b312f4de7485
Depends-On: Ic990ab108afc2791f9db297a07cfe1bdc59dbbf2
Depends-On: I434a8ac89f21134bb5c0e139fea08ebf66a83437
Change-Id: Ibc35c880fe872e168a2fc290b89cad7db70e2228