tests: move slow ScopeStructureTest out of structure
While looking at the list of tests for an extension I found 3000+
ScopeStructureTest which are generated from the php files in
mediawiki/core (more precisely `$wgAutoloadLocalClasses`):
* those tests take 21 seconds to complete on my machine.
* None were generated for the extension being tested, those tests are
  thus solely affected by mediawiki/core.

`tests/phpunit/structure` is included in the `extensions` and `skins`
PHPUnit testsuites and any patches made to them would run that 21
seconds suite even though its only testing mediawiki/core.

Move the test outside of `structure` so it is no longer run for
`extensions` and `skins`.

Bug: T225730
Change-Id: I628210b8b270773f3dad12bbde9d72f0328fcceb