PageEditStash: forward-compatibility with a future cache key version bump
This will allow us to cleanly bump the cache key for the next release
(I42e1303236318862ecae856e2f9207e3778764f5) without dropping stash
entries when this is rolled out on the train.

At the moment this only adds the forward-compatibility piece without
changing the cache keys for the versions being currently used.

Change-Id: Iefffff054ff4fdec10169fec287e5493d8b4af7d