RecipientSet: Add removeRecipient() method
We have addRecipient(), this seems like an obvious companion,
and I need it to implement middleware in Echo.

In order to make removing a recipient a constant-time operation,
convert the storage to be an associative array. UserIdentity objects
are compared by the user name, so we can use them as the array keys.

As a bonus, this gives us deduplication, making this set really a set.

Change-Id: I02ff1c2cf469354016140a00e2c640346ce7bb12