selenium: Use ECMAScript modules in RunJobs.js
RunJobs.js was one of the rare files that didn't need to be converted to
ECMAScript modules (ESM) for core Selenium tests to pass.
According to Codesearch, the file is used in at least PageTriage and MinervaNeue.
This commit converts the file to ESM so it could be used in mentioned repositories.

Bug: T398046
Change-Id: I5cc9091f763ae2d2d363b5540f00f3ea83eeedea