search: create page in the searched namespace
When a single namespace is selected in search and we are offering
to create a page with the matching name, use the selected
namespace as part of the title.

One awkward edge case is that this still follows standard title
parsing rules. So if the search string is 'User:Example' but the
'Category' namespace is selected, the 'User' namespace in the
string will still take precedence.

Bug: T398133
Change-Id: Ib2166a2e2ad58985537064f7deee4b96ae2a2326