typeahead: Add hook to augment api parameters
As part of AB testing changes to search we need the ability to
augment the URL being used for typeahead to apply a test
treatment.  Add an mw.hook that allows for this use case.

Note that this is still only best effort, extensions have the
ability to replace the typeahead implementation which would
skip all of this code. The only current implementation of that I
could find is in wikibase, meaning this will not work on
wikis that allow wikibase to replace the typeahead.

Bug: T397732
Change-Id: I1a477dfaa2b41846cb0167d6639e14b2abbe1e33
(cherry picked from commit 2e6723e61cad6393d8d104e4b0ac2f3d212d4551)