selenium: Throw exception when API response lacks expected field
This code was added in Ic82e9849e6aa3d867, where it was agreed that we
didn't need defensive coding because the response is guaranteed to
contain the `groups` property. However, judging from recent errors in
beta, that doesn't seem to be the case.

So, check whether the property exists and throw an exception if it
doesn't, printing the entire response so the issue can hopefully be
debugged.

Bug: T393428
Change-Id: Ib3327b786ee5b0fc1ec35e5afedf3f5d64d85f85