Document Status and StatusValue as generic classes with `@template`
These annotations allow IDEs and static analysis tools to better
understand the types in code such as:

    /** @return StatusValue<Foo> */
    function doStuff(): StatusValue { ... }
    doStuff()->getValue()->bar();

Phan will now validate that the class 'Foo' has a method called 'bar',
and PhpStorm will autocomplete its methods after `...->getValue()->`.

As of this commit, the only way to make Phan type check the generic
type parameter is to document it in PHPDoc. Changes to make it infer
the generic type of a Status or StatusValue when constructing it
are left for another commit to make things more understandable.

Extension compatibility note: Extensions inheriting from StatusValue
must add `@extends StatusValue<mixed>` in the class PHPDoc to avoid
Phan errors (or a more specific type; and likewise for Status). Also,
extensions that already used `StatusValue<Foo>` etc. in their PHPDocs
will now have it validated by Phan, which may result in errors.

Changes to other files:
* Document specialized Status/StatusValue subclasses and their types
* Update Phan suppressions (I'm not sure what causes these changes.
  It's probably some edge cases where the `mixed` type, which was
  previously documented for StatusValue::$value, and the empty union
  type, which is what you get substituted for `T` when you don't
  provide the type parameter, are analyzed slightly differently,
  even though they should be pretty much the same.)

Bug: T397781
Change-Id: I76dacc5809aeefe6f287a227e28e3a57ccf01d67